Move BIDSGrabber to interfaces.io
