+++ nipype/interfaces/io.py
@@ -31,6 +31,7 @@ import shutil
 import subprocess
 import re
 import tempfile
+from os.path import join, dirname
 from warnings import warn
 
 import sqlite3
@@ -38,10 +39,15 @@ import sqlite3
 from .. import config, logging
 from ..utils.filemanip import copyfile, list_to_filename, filename_to_list
 from ..utils.misc import human_order_sorted, str2bool
-from .base import (TraitedSpec, traits, Str, File, Directory, BaseInterface,
-                   InputMultiPath, isdefined, OutputMultiPath,
-                   DynamicTraitedSpec, Undefined, BaseInterfaceInputSpec)
-from .bids_utils import BIDSDataGrabber
+from .base import (
+    TraitedSpec, traits, Str, File, Directory, BaseInterface, InputMultiPath,
+    isdefined, OutputMultiPath, DynamicTraitedSpec, Undefined, BaseInterfaceInputSpec)
+
+have_pybids = True
+try:
+    from bids import grabbids as gb
+except ImportError:
+    have_pybids = False
 
 try:
     import pyxnat
@@ -2716,4 +2722,122 @@ class JSONFileSink(IOBase):
 
         outputs = self.output_spec().get()
         outputs['out_file'] = out_file
+        return outputs
+
+
+class BIDSDataGrabberInputSpec(DynamicTraitedSpec):
+    base_dir = Directory(exists=True,
+                         desc='Path to BIDS Directory.',
+                         mandatory=True)
+    output_query = traits.Dict(key_trait=Str,
+                               value_trait=traits.Dict,
+                               desc='Queries for outfield outputs')
+    raise_on_empty = traits.Bool(True, usedefault=True,
+                                 desc='Generate exception if list is empty '
+                                 'for a given field')
+    return_type = traits.Enum('file', 'namedtuple', usedefault=True)
+
+
+class BIDSDataGrabber(IOBase):
+
+    """ BIDS datagrabber module that wraps around pybids to allow arbitrary
+    querying of BIDS datasets.
+
+    Examples
+    --------
+
+    By default, the BIDSDataGrabber fetches anatomical and functional images
+    from a project, and makes BIDS entities (e.g. subject) available for
+    filtering outputs.
+
+    >>> bg = BIDSDataGrabber()
+    >>> bg.inputs.base_dir = 'ds005/'
+    >>> bg.inputs.subject = '01'
+    >>> results = bg.run() # doctest: +SKIP
+
+
+    Dynamically created, user-defined output fields can also be defined to
+    return different types of outputs from the same project. All outputs
+    are filtered on common entities, which can be explicitly defined as
+    infields.
+
+    >>> bg = BIDSDataGrabber(infields = ['subject'], outfields = ['dwi'])
+    >>> bg.inputs.base_dir = 'ds005/'
+    >>> bg.inputs.subject = '01'
+    >>> bg.inputs.output_query['dwi'] = dict(modality='dwi')
+    >>> results = bg.run() # doctest: +SKIP
+
+    """
+    input_spec = BIDSDataGrabberInputSpec
+    output_spec = DynamicTraitedSpec
+    _always_run = True
+
+    def __init__(self, infields=None, outfields=None, **kwargs):
+        """
+        Parameters
+        ----------
+        infields : list of str
+            Indicates the input fields to be dynamically created
+
+        outfields: list of str
+            Indicates output fields to be dynamically created.
+            If no matching items, returns Undefined.
+        """
+        super(BIDSDataGrabber, self).__init__(**kwargs)
+
+        if not isdefined(self.inputs.output_query):
+            self.inputs.output_query = {"func": {"modality": "func"},
+                                        "anat": {"modality": "anat"}}
+
+        # If infields is empty, use all BIDS entities
+        if infields is None and have_pybids:
+            bids_config = join(dirname(gb.__file__), 'config', 'bids.json')
+            bids_config = json.load(open(bids_config, 'r'))
+            infields = [i['name'] for i in bids_config['entities']]
+        # if outfields is not set, return those defined in infields
+        if infields and outfields is None:
+            outfields = [infield for infield in infields]
+
+        self._infields = infields or []
+        self._outfields = outfields or []
+
+        # used for mandatory inputs check
+        undefined_traits = {}
+        for key in self._infields:
+            self.inputs.add_trait(key, traits.Any)
+            undefined_traits[key] = kwargs[key] if key in kwargs else Undefined
+
+        self.inputs.trait_set(trait_change_notify=False, **undefined_traits)
+
+    def _run_interface(self, runtime):
+        if not have_pybids:
+            raise ImportError(
+                "The BIDSEventsGrabber interface requires pybids."
+                " Please make sure it is installed.")
+        return runtime
+
+    def _list_outputs(self):
+        layout = gb.BIDSLayout(self.inputs.base_dir)
+
+        # If infield is not given nm input value, silently ignore
+        filters = {}
+        for key in self._infields:
+            value = getattr(self.inputs, key)
+            if isdefined(value):
+                filters[key] = value
+
+        outputs = {}
+        for key, query in self.inputs.output_query.items():
+            args = query.copy()
+            args.update(filters)
+            filelist = layout.get(return_type=self.inputs.return_type, **args)
+            if len(filelist) == 0:
+                msg = 'Output key: %s returned no files' % key
+                if self.inputs.raise_on_empty:
+                    raise IOError(msg)
+                else:
+                    iflogger.warning(msg)
+                    filelist = Undefined
+
+            outputs[key] = filelist
         return outputs
